requirements.txt
fastapi
uvicorn
python-docx
PyMuPDF
requests
openai
pinecone
python-dotenv
google-generativeai